Use phonenumbers for timezone lookup except in US and AU (#172)
* Use phonenumbers for timezone lookup except in US and AU

This is so much more accurate for "rest of the world".

For Canada we were missing some new area codes, it makes
sense to offload to phonenumbers as well.

For US and AU we can make follow up commits to migrate as well
but we might be more careful there and tackle any edge cases.

* Support for python 3.8
* Make data loading more mypy friendly